Create some changes in the test where it was added the permission change to the ping.sh script to make it executable. Also, for testing it was comment the ubuntu OS.
Adding a value to x in the ping.sh script and removing the comments to ubuntu OS

Adding to save the test 1 script output

Fixing the test 1 ping.sh script.

Fixing the script to test and to stop infinity loop.

Fixing the script to increment the variable.

Fixing increment of variable x.

Fixing and modifying the ping script to make the script more efficient.

Adding more description to the ping script output

Adding to copy the python script and adding to bash script to be deleted on cleaning.

Deleting the test file for testing.

Changing the length of the script from 25 to 40. activating the delete vlans and cleaning roles

Fixing a bug on the python script to delete the evcs and any empty keys in the dict.

Passed the roles to the local playbook.

Adding more description fro the tasks in ansible

Testing the description of tasks in hosts.

Testing a way to skip the localhost from the tasks

Testing a way to skip the localhost from the tasks and installing python before to run the creation of evc

Testing to see if it display the other hosts

Testing to see if it display the other hosts

Testing to see if it display the other hosts by comment the localhost in inventory and creating the connection inside the playbook local.

Changing to a delegate to localhost.

there is an error for hosts. trying to find a solution

Comment the localhost tasks to check if it fix the problem

Added the localhost to the inventory to check if it would fix the error.

importing the playbook for tests to eliminate the error

Adding the name to fix an error.

Changing the extension to yaml

Importing the the playbook as a tasks

Leaving all as it was before.

Activating all roles and cleaning the code

Comment some tasks to test why is freeze

Comment the task to test

Fixing a bug in the creation and deletion of the EVCs.

Comment the deletion of the key